prepare cleanup of daemon config
